requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
mysql-connector-python==8.2.0
pymysql==1.1.0
cryptography==41.0.7
openpyxl==3.1.2
pandas==2.2.3
reportlab==4.0.7
python-dotenv==1.0.0
email-validator==2.3.0
Werkzeug==3.0.1
gunicorn==21.2.0
